fix timezone support in wp2zinnia management command
In the WP export, 'post_date' represents the timezone indicated by the server, and 'post_date_gmt' represents GMT (or possibly UTC) for the same time.  On import, the management command is currently using the local timestamp but UTC as the timezone with which to interpret it, making all the times off by the timezone offset.

This solution (for entries and comments) uses the gmt values and 'GMT' as the timezone when saving datetimes.