More generic "extensions" parameter for Markdown
In the current state, users can only provide a comma-separated list of extension paths. This prevents the users from givin extension-specific settings such as `TocExtension(marker="--TOC--", permalink=True)`

In this patch we only do the splitting if the extension parameter is a string (using `six.string_types` of course) so the caller can use a list of Extension instances instead.