Calculate unicode punctuation dictionary lazily
Whilst profiling my app I found that this loop over all unicode characters was
taking up to half a second of startup time. This is a high tax for just
installing Zinnia, since it affects every test run etc.

A quick timeit shows it takes 386ms - however it's slower in our dev
environment which is in a VM in Vagrant:

```
In [7]: %timeit PUNCTUATION = dict.fromkeys(i for i in range(sys.maxunicode) if unicodedata.category(six.unichr(i)).startswith('P') )
1 loops, best of 3: 386 ms per loop
```

It can be calculated once on-demand, obviating the need for the high
computation cost to be paid at import time. However this does mean that the
first time it's called in a process it will be slow.

An alternative would be to install the `regex` library which can replace
unicode characters by their class (see
http://stackoverflow.com/a/11066687/1427135) - however I wasn't sure if adding
another dependency would be worth it.